requirements.txt
# Runtime dependency baseline (as-of March 2026)
# Python >= 3.11

# Core rendering and image processing
opencv-python>=4.13.0.92,<4.14
Pillow>=12.1.1,<12.2
numpy>=2.4.4,<2.5
scipy>=1.17.1,<1.18
freetype-py>=2.5.1,<2.6

# Document degradation and desktop telemetry
augraphy>=8.2.6,<8.3
pyautogui>=0.9.54,<0.10

# OCR benchmarking utilities
# EasyOCR pulls Torch as a transitive dependency.
easyocr>=1.7.2,<1.8

# Optional visualization/debugging
matplotlib>=3.10.8,<3.11

# Dev/test
pytest>=9.0.2,<9.1
